copy-builtin: remove .gitignore from KERNEL_DIR/include/zfs
While it makes sense to make sure we don't control the generated
zfs_gitrev.h in the zfs source tree, once we copy the module sources into a
kernel source tree that's a problem.  If you're keeping your zfs sources in
a kernel tree on a branch, you want to commit all the required files.