added zfs-auto-snapshot as a submodule
The autogen.sh script now initializes submodules and the contrib Makefile
installs zfs-auto-snapshot in our configured prefix (but does NOT install
its cron files).

Also, the toplevel Makefile was removing all Makefiles recursively in a
distclean rule... which shouldn't be needed (automake's Makefile already
does this) and was removing the non-automake Makefile in zfs-auto-snapshot.