copy-builtin: remove .gitignore from KERNEL_DIR/fs/zfs
While it makes sense to make sure we don't control generated files in
the zfs source tree, once we copy the module sources into a kernel
source tree that's a problem.  If you're keeping your zfs sources in a
kernel tree on a branch, you want to commit all the required files.

NOTE: This version of the commit uses '-f' so it won't fail when run
      from an extracted dist tarball.

Signed-off-by: Michael D Labriola <michael.d.labriola@gmail.com>